Restructured files to make sense
